+++ src/Provider/Factory.php
@@ -10,7 +10,7 @@ declare(strict_types = 1);
 namespace Dot\Navigation\Provider;
 
 use Dot\Navigation\Exception\RuntimeException;
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 
 /**
  * Class Factory