+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## Unreleased(0.2.0)
+
+Migration to ZE2
+
+### Changed
+* updated dependencies
+* updated to PSR11 container
+* updated middleware to PSR15
+
+### Added
+* Nothing
+
+### Deprecated
+* Nothing
+
+### Removed
+* Nothing
+
+### Fixed
+* Nothing
+
+
 ## 0.1.1 - 2017-03-11
 
 Updated php file headers doc blocks